requirements.txt
streamlit>=1.35
plotly
matplotlib
pandas
pymongo>=4.0
certifi
pdfplumber
openpyxl
reportlab
python-dotenv
httpx
requests
pytest
pytest-mock
ruff
black
openai
huggingface_hub
google-generativeai
yfinance
bcrypt
supabase
